apple II/virtu - some reorg
