+++ BizHawk.Emulation.Cores/BizHawk.Emulation.Cores.csproj
@@ -50,12 +50,8 @@
 		<PackageReference Include="ELFSharp" Version="0.1.0" PrivateAssets="All" />
 		<PackageReference Include="Newtonsoft.Json" Version="12.0.3" PrivateAssets="All" />
 		<PackageReference Include="OpenTK" Version="3.0.1" PrivateAssets="All" />
-		<Reference Include="PeNet"
-			HintPath="$(SolutionDir)References/PeNet.dll"
-			Private="true" />
-		<Reference Include="Virtu"
-			HintPath="$(SolutionDir)References/Virtu.dll"
-			Private="true" />
+		<Reference Include="PeNet" HintPath="$(SolutionDir)References/PeNet.dll" Private="true" />
+		<Reference Include="Virtu" HintPath="$(SolutionDir)References/Virtu.dll" Private="true" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="$(SolutionDir)BizHawk.BizInvoke/BizHawk.BizInvoke.csproj" />
