+++ BizHawk.Emulation.Cores/Computers/AppleII/Components/CassetteComponent.cs
