+++ BizHawk.Emulation.Cores/Computers/AppleII/Components/GamePortComponent.cs
